Backend/requirements.txt
Flask
google-generativeai
python-dotenv
flask-restx